mailstic/requirements.txt
fastapi[standard]>=0.115.4
alembic==1.13.3
markdown==3.7
SQLAlchemy==2.0.36
SQLalchemy[asyncio]==2.0.36
sqlmodel==0.0.22
aiomysql==0.2.0
PyMySql==1.1.1
streamlit==1.40.0
numpy==1.26.4
pandas==2.2.3
streamlit-antd-components==0.3.2
opencv-python==4.10.0.84
plotly==5.24.1
pymorphy3==2.0.2
gensim==4.3.3
scipy==1.13.1
paddlepaddle==2.6.2
paddleocr==2.9.1
pillow==11.0.0
matplotlib==3.9.2
fuzzywuzzy==0.18.0
catboost==1.2.5
scikit-learn==1.5.2
watchdog
setuptools==75.2.0
python-slugify==8.0.4
torch==2.3.1
transformers==4.41.2